requirements.txt
adal==1.2.7
adlfs==2023.10.0
aiobotocore==2.5.0
aiofiles==23.1.0
aiohttp==3.8.6
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.12.1
amqp==5.1.1
annotated-types==0.6.0
anyio==4.0.0
apache-airflow==2.7.2
apache-airflow-providers-amazon==8.0.0
apache-airflow-providers-celery==3.1.0
apache-airflow-providers-cncf-kubernetes==6.1.0
apache-airflow-providers-common-sql==1.8.0
apache-airflow-providers-docker==3.6.0
apache-airflow-providers-elasticsearch==4.4.0
apache-airflow-providers-ftp==3.6.0
apache-airflow-providers-google==10.0.0
apache-airflow-providers-grpc==3.1.0
apache-airflow-providers-hashicorp==3.3.1
apache-airflow-providers-http==4.6.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-microsoft-azure==6.0.0
apache-airflow-providers-mysql==5.0.0
apache-airflow-providers-odbc==3.2.1
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-redis==3.1.0
apache-airflow-providers-sendgrid==3.1.0
apache-airflow-providers-sftp==4.2.4
apache-airflow-providers-slack==7.2.0
apache-airflow-providers-snowflake==4.0.5
apache-airflow-providers-sqlite==3.5.0
apache-airflow-providers-ssh==3.6.0
apispec==6.3.0
argcomplete==3.1.3
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.0
azure-batch==13.0.0
azure-common==1.1.28
azure-core==1.29.5
azure-cosmos==4.3.1
azure-datalake-store==0.0.53
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-kusto-data==0.0.45
azure-mgmt-containerinstance==1.5.0
azure-mgmt-core==1.4.0
azure-mgmt-datafactory==1.1.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==23.0.0
azure-nspkg==3.0.2
azure-servicebus==7.9.0
azure-storage-blob==12.18.3
azure-storage-common==2.1.0
azure-storage-file==2.1.0
azure-storage-file-datalake==12.11.0
azure-synapse-spark==0.7.0
Babel==2.13.1
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
blinker==1.6.3
boto3==1.26.76
botocore==1.29.76
bs4==0.0.1
cachelib==0.9.0
cachetools==5.3.0
cattrs==23.1.2
celery==5.2.7
certifi==2023.7.22
cffi==1.16.0
chardet==5.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
cloudpickle==2.2.1
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.5
dask==2023.4.1
db-dtypes==1.1.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.6
distributed==2023.4.1
dnspython==2.4.2
docker==6.0.1
docutils==0.20.1
elasticsearch==7.13.4
elasticsearch-dbapi==0.2.10
elasticsearch-dsl==7.4.1
email-validator==1.3.1
eventlet==0.33.3
exceptiongroup==1.1.1
filelock==3.12.0
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flower==1.2.0
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gcloud-aio-auth==4.2.1
gcloud-aio-bigquery==6.3.0
gcloud-aio-storage==8.2.0
geographiclib==2.0
geopy==2.4.0
gevent==22.10.2
google-api-core==2.8.2
google-api-python-client==1.12.11
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-aiplatform==1.16.1
google-cloud-appengine-logging==1.1.3
google-cloud-audit-log==0.2.4
google-cloud-automl==2.8.0
google-cloud-bigquery==2.34.4
google-cloud-bigquery-datatransfer==3.7.0
google-cloud-bigquery-storage==2.14.1
google-cloud-bigtable==2.11.1
google-cloud-build==3.9.0
google-cloud-compute==0.7.0
google-cloud-container==2.11.1
google-cloud-core==2.3.2
google-cloud-datacatalog==3.9.0
google-cloud-dataflow-client==0.5.4
google-cloud-dataform==0.2.0
google-cloud-dataplex==1.1.0
google-cloud-dataproc==5.0.0
google-cloud-dataproc-metastore==1.6.0
google-cloud-dlp==3.8.0
google-cloud-kms==2.12.0
google-cloud-language==1.3.2
google-cloud-logging==3.2.1
google-cloud-memcache==1.4.1
google-cloud-monitoring==2.11.0
google-cloud-orchestration-airflow==1.4.1
google-cloud-os-login==2.7.1
google-cloud-pubsub==2.13.5
google-cloud-redis==2.9.0
google-cloud-resource-manager==1.6.0
google-cloud-secret-manager==1.0.2
google-cloud-spanner==1.19.3
google-cloud-speech==1.3.4
google-cloud-storage==2.8.0
google-cloud-tasks==2.10.1
google-cloud-texttospeech==1.0.3
google-cloud-translate==1.7.2
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-cloud-workflows==1.7.1
google-crc32c==1.5.0
google-re2==1.1
google-resumable-media==2.5.0
googleapis-common-protos==1.61.0
graphviz==0.20.1
greenlet==3.0.1
grpc-google-iam-v1==0.12.4
grpcio==1.59.2
grpcio-gcp==0.2.2
grpcio-status==1.48.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httplib2==0.21.0
httpx==0.25.0
humanize==4.6.0
hvac==1.1.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
json-merge-patch==0.2
jsonpath-ng==1.5.3
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kombu==5.2.4
kubernetes==23.6.0
kubernetes-asyncio==24.2.2
lazy-object-proxy==1.9.0
ldap3==2.9.1
limits==3.6.0
linkify-it-py==2.0.2
locket==1.0.0
lockfile==0.12.2
looker-sdk==23.6.0
lxml==4.9.2
Mako==1.2.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
msal==1.25.0
msal-extensions==1.0.0
msgpack==1.0.5
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.4
mypy-boto3-appflow==1.26.123
mypy-boto3-rds==1.26.116
mypy-boto3-redshift-data==1.26.109
mysqlclient==2.1.1
numpy==1.26.1
oauthlib==3.2.2
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
ordered-set==4.1.0
oscrypto==1.3.0
packaging==23.2
pandas==2.1.2
pandas-gbq==0.17.9
paramiko==3.1.0
partd==1.4.0
pathspec==0.11.2
pendulum==2.1.2
platformdirs==3.5.0
pluggy==1.3.0
ply==3.11
portalocker==2.8.2
prison==0.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.19.6
protobuf==4.24.4
psutil==5.9.6
psycopg2-binary==2.9.6
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.17
pydantic==2.4.2
pydantic_core==2.10.1
pydata-google-auth==1.7.0
Pygments==2.16.1
pyhcl==0.4.4
PyJWT==2.8.0
PyNaCl==1.5.0
pyodbc==4.0.39
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-ldap==3.4.3
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
redis==3.5.3
redshift-connector==2.0.910
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.6.0
rich-argparse==1.4.0
rpds-py==0.10.6
rsa==4.9
s3transfer==0.6.0
scramp==1.4.4
sendgrid==6.10.0
setproctitle==1.3.3
six==1.16.0
slack-sdk==3.21.3
sniffio==1.3.0
snowflake-connector-python==3.0.3
snowflake-sqlalchemy==1.4.7
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.50
sqlalchemy-bigquery==1.6.1
SQLAlchemy-JSONField==1.0.1.post0
sqlalchemy-redshift==0.8.14
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
sshtunnel==0.4.0
starkbank-ecdsa==2.2.0
statsd==4.0.1
tabulate==0.9.0
tblib==1.7.0
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
toolz==0.12.0
tornado==6.3.1
typing_extensions==4.8.0
tzdata==2023.3
uamqp==1.6.4
uc-micro-py==1.0.2
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==2.0.7
vine==5.0.0
virtualenv==20.23.0
watchtower==2.0.1
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.1.1
yarl==1.9.2
zict==3.0.0
zipp==3.17.0
zope.event==4.6
zope.interface==6.0
